refactor(shared-types): move preview types under types
