+++ src/shared/types/excelPreview.ts
