+++ src/shared/types/wordPreview.ts
